Refactor file structure of models module
